some fix

+++ 0x0A-argc_argv/4-add.c
@@ -23,7 +23,7 @@ int main(int argc, char *argv[])
 
 	for (i = 1; i < argc; i++)
 	{
-		for (j = 0; j < strlen(argv[i]); j++)
+		for (j = 0; j < (int) strlen(argv[i]); j++)
 		{
 			if (!((int) argv[i][j] >= 48 && (int) argv[i][j] <= 57))
 			{
